requirements.txt
fontTools
Pillow
